Día 6
Modificaciones en el arbol y en la toma de tiempos

+++ [EDD]Practica2_201020331/Nodo_arbol.h
@@ -2,7 +2,7 @@
 #define NODO_ARBOL_H_INCLUDED
 
 typedef struct node{
-    int iden;
+    long int iden;
     int fe;
     struct node *left_son;
     struct nodo *right_son;
