+++ [EDD]Practica2_201020331/Nodo_lista.h
@@ -2,8 +2,8 @@
 #define NODO_LISTA_H_INCLUDED
 
  typedef struct Nodo_lista{
-    int clave;
-    int pos;
+    long int clave;
+    long int pos;
     int stop;
     struct Nodo_lista *next;
     struct Nodo_lista *bak;
